Move the `RunLengthStream` from `src/core/stream.js` and into its own file
